Move login components to page folder
