+++ src/pages/Login/Login.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import { LoginComponent } from '../../components/LoginComponent';
-import { RegistrationComponent } from '../../components/RegistrationComponent';
+import { LoginComponent } from './components/LoginComponent';
+import { RegistrationComponent } from './components/RegistrationComponent';
 import { Star } from '../../components/Start';
 
 import styles from './Login.css';
