+++ src/pages/Login/components/RegistrationComponent/RegistrationComponent.css
