+++ src/pages/Login/components/RegistrationComponent/index.ts
