+++ src/pages/Login/components/RegistrationComponent/types.ts
